Move dict_path function to __init__.py
